Moved teamData from pages to component
